rename inp module to model
