+++ honeycomb-kernels/src/grisubal/model.rs
